réecriture des politiques mis chemin

+++ app/Http/Controllers/Bordereau/CommercialController.php
@@ -14,54 +14,59 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
+use Illuminate\Support\Facades\Gate;
 
 class CommercialController extends Controller
 {
     public function all(): JsonResponse
     {
-        $commerciaux = Commercial::select('id', 'code', 'created_at', 'site_id', 'user_id')->with('site:id,nom')
-            ->with(['user' => fn($query) => $query->select('id', 'name')->without('avatar')])->get();
+        $response = Gate::inspect('viewAny', Commercial::class);
+        $query = Commercial::select('id', 'code', 'created_at', 'site_id', 'user_id')->with('site:id,nom', 'user:id,name');
+        $commerciaux = $response->allowed() ? $query->get() : $query->owner()->get();
         return response()->json(['commerciaux' => CommercialListResource::collection($commerciaux)]);
     }
 
     public function getPaginate(): JsonResource
     {
-        $commerciaux = Commercial::select('id', 'code', 'created_at', 'site_id', 'user_id')->with('site:id,nom')
-            ->with(['user' => fn($query) => $query->select('id', 'name')->without('avatar')])->paginate(10);
+        $response = Gate::inspect('viewAny', Commercial::class);
+        $query = Commercial::select('id', 'code', 'created_at', 'site_id', 'user_id')->with('site:id,nom', 'user:id,name');
+        $commerciaux = $response->allowed() ? $query->paginate(10) : $query->owner()->paginate(10);
         return CommercialListResource::collection($commerciaux);
     }
 
     public function getSearch(string $search): JsonResource
     {
-        $commerciaux = Commercial::select('id', 'code', 'created_at', 'site_id', 'user_id')->with('site:id,nom')
-            ->with(['user' => fn($query) => $query->select('id', 'name')->without('avatar')])
+        $response = Gate::inspect('viewAny', Commercial::class);
+        $query = Commercial::select('id', 'code', 'created_at', 'site_id', 'user_id')->with('site:id,nom', 'user:id,name')
             ->whereRaw("DATE_FORMAT(commercials.created_at,'%d-%m-%Y') LIKE ?", "$search%")->orWhere('code', 'LIKE', "%$search%")
-            ->orWhereHas('user', fn(Builder $query): Builder => $query->where('name', 'LIKE', "%$search%"))->paginate(10);
+            ->orWhereHas('user', fn(Builder $query): Builder => $query->where('name', 'LIKE', "%$search%"));
+        $commerciaux = $response->allowed() ? $query->paginate(10) : $query->owner()->paginate(10);
         return CommercialListResource::collection($commerciaux);
     }
 
     public function getSelect(): JsonResponse
     {
-        $commerciaux = Commercial::select('id', 'code', 'user_id')->with([
-            'user' => fn($query) => $query->select('id', 'name')->without('avatar'),
-        ])->get();
+        $response = Gate::inspect('viewAny', Commercial::class);
+        $query = Commercial::select('id', 'code', 'user_id')->with('user:id,name');
+        $commerciaux = $response->allowed() ? $query->get() : $query->owner()->get();
         return response()->json(['commerciaux' => CommercialSelectResource::collection($commerciaux)]);
     }
 
     public function store(Request $request): JsonResponse
     {
+        $this->authorize('create', Commercial::class);
         $request->validate(Commercial::RULES);
         $commercial = new Commercial($request->all());
         $commercial->codeGenerate();
         $commercial->save();
-        $message = "Le commercial: $commercial->code a été enregistré avec succès.";
-        return response()->json(['message' => $message]);
+        return response()->json(['message' => "Le commercial: $commercial->code a été enregistré avec succès."]);
     }
 
     public function attribuer(AssignationRequest $request): JsonResponse
     {
-        $bordereau = Bordereau::make($request->validated());
         $commercial = Commercial::with('user:id,name')->find($request->commercial_id);
+        $this->authorize('attribuate', $commercial);
+        $bordereau = Bordereau::make($request->validated());
         $bordereau->codeGenerate();
         $bordereau->site_id = $commercial->site_id;
         $bordereau->save();
@@ -73,6 +78,7 @@ class CommercialController extends Controller
     public function trash(int $id): JsonResponse
     {
         $commercial = Commercial::find($id);
+        $this->authorize('delete', $commercial);
         $commercial->delete();
         $message = "Le commercial: $commercial->code a été supprimé avec succès.";
         return response()->json(['message' => $message]);
@@ -81,6 +87,7 @@ class CommercialController extends Controller
     public function restore(int $id): JsonResponse
     {
         $commercial = Commercial::withTrashed()->find($id);
+        $this->authorize('restore', $commercial);
         $commercial->restore();
         $message = "Le commercial $commercial->code a été restauré avec succès.";
         return response()->json(['message' => $message]);
@@ -88,19 +95,25 @@ class CommercialController extends Controller
 
     public function trashed(): JsonResponse
     {
-        $commerciaux = Commercial::onlyTrashed()->get();
+        $response = Gate::inspect('viewAny', Commercial::class);
+        $this->authorize('viewAny', Commercial::class);
+        $query = Commercial::select('id', 'code', 'created_at', 'site_id', 'user_id')->with('site:id,nom', 'user:id,name')->onlyTrashed();
+        $commerciaux = $response->allowed() ? $query->get() : $query->owner()->get();
         return response()->json(['commerciaux' => $commerciaux]);
     }
 
     public function show(int $id): JsonResponse
     {
         $commercial = Commercial::with('user:id,name', 'site:id,nom')->find($id);
+        $this->authorize('view', $commercial);
         return response()->json(['commercial' => CommercialResource::make($commercial)]);
     }
 
     public function getMonthBordereaux(Commercial $commercial): JsonResource
     {
-        $bordereaux = $commercial->bordereaux()->whereYear('jour', now()->year)->whereMonth('jour', now()->month)->get();
+        $response = Gate::inspect('viewAny', Bordereau::class);
+        $query = $commercial->bordereaux()->whereYear('jour', now()->year)->whereMonth('jour', now()->month);
+        $bordereaux = $response()->allowed() ? $query->get() : $query->owner()->get();
         return BordereauResource::collection($bordereaux);
     }
 }
